refactoring and restructuring
